% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFCRE.R
\name{pfcre}
\alias{pfcre}
\title{pfcre}
\usage{
pfcre(
  formula = NULL,
  addcovars = NULL,
  data = NULL,
  id = NULL,
  family = NULL,
  intercept = TRUE,
  weights = rep(1, nobs),
  initial.beta,
  mustart,
  eta.new,
  lambda = NULL,
  max.steps = 1000,
  degree = 2
)
}
\arguments{
\item{formula}{an object of class \code{\link{formula}}: "y~x1+x2+x3..."}

\item{addcovars}{an object of class \code{\link{formula}}: "~z1+z2+z3..."}

\item{data}{an object of class \code{\link{data.frame}} containing the variables to be
used in the model, including group ids}

\item{id}{a variable in \code{data} that identifies groups}

\item{family}{a descriptions of the error distribution and link function to be used in the
model. See \code{\link{family}} for details of family functions.}

\item{weights}{an optional vector of prior weights to be used in the fitting process.}

\item{initial.beta}{starting values for the parameters in the linear predictor (optional)}

\item{mustart}{starting values for the vector of means (optional)}

\item{eta.new}{starting values for the linear predictor (optional)}

\item{lambda}{is the parameter value for the \code{scadpen} penalty. "a" is set to 3.7}
}
\value{
Returns an object with the following elements:
\item{beta}{ a named vector of estimated main coefficients}
\item{gamma}{ a named vector of estimated coefficients for the polynomial}
\item{sdeta}{ the variance of the random effect}
\item{sdbeta}{ a named vector of the standard error of the \code{beta} coefficients}
\item{sdgamma}{ a named vector of the standard error of the \code{gamma} coefficients}
\item{vcov}{ a matrix of the covariance of \code{beta} and \code{gamma}}
\item{nobs}{ the number of observations}
\item{loglik}{ the unpenalized log-likelihood}
\item{formula}{ the formula used in the call}
}
\description{
\code{pfcre} implements the Penalized Flexible Correlated Random Effects estimator
}
