% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFCREms.R
\name{pfcrems}
\alias{pfcrems}
\title{pfcrems}
\usage{
pfcrems(
  formula = NULL,
  addcovars = NULL,
  data = NULL,
  id = NULL,
  family = NULL,
  intercept = TRUE,
  weights = rep(1, nobs),
  initial.beta,
  mustart,
  eta.new,
  lambda.cand = NULL,
  max.steps = 1000,
  degree = 2
)
}
\arguments{
\item{formula}{an object of class \code{\link{formula}}: "y~x1+x2+x3..."}

\item{addcovars}{an object of class \code{\link{formula}}: "~z1+z2+z3..."}

\item{data}{an object of class \code{\link{data.frame}} containing the variables to be
used in the model, including group ids}

\item{id}{a variable in \code{data} that identifies groups}

\item{family}{a descriptions of the error distribution and link function to be used in the
model. See \code{\link{family}} for details of family functions.}

\item{weights}{vector of prior weights to be used in the fitting process (optional)}

\item{initial.beta}{starting values for the parameters in the linear predictor (optional)}

\item{mustart}{starting values for the vector of means (optional)}

\item{eta.new}{starting values for the linear predictor (optional)}

\item{lambda.cand}{is a vector of candidate values for the \code{scadpen} penalty. "a" is set to 3.7}
}
\value{
Returns an object with the following elements:
\item{Model}{ a list with the estimates from the best model selected via Akaike Information Criterion}
\item{lambda.cand}{ a vector with the original candidate values for the \code{scadpen} penalty}
\item{cv}{ a vector with the model fit for the \code{labmda.cand} elements}
}
\description{
\code{pfcrems} performs model penalty selection for the Penalized Flexible Correlated Random Effects estimator
}
